Organise file structure. Adjust imports and remove index.css. Adjust app.css. Closes #34
