+++ client/src/main.tsx
@@ -1,10 +1,10 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
-import Login from './components/Login.tsx'
-import Register from './components/Register.tsx'
-import App from './App.tsx'
+
+import Login from './routes/LoginScreen.tsx'
+import Register from './routes/RegisterScreen.tsx'
+import App from './routes/App.tsx'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
